chainfly-backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sqlalchemy
jinja2
aiofiles
apscheduler
python-dotenv 